- Send server sync packets on every frame. - Add smoothing to player RTT. Increase server offset smoothing.

+++ PolyPlane/Net/NetHost/ServerNetHost.cs
@@ -117,7 +117,7 @@ namespace PolyPlane.Net.NetHost
             }
         }
 
-        public override uint GetPlayerRTT(int playerID)
+        public override double GetPlayerRTT(int playerID)
         {
             if (_peers.TryGetValue((uint)playerID, out var peer))
             {
